Refined HTML and CSS
I made the first image outside of the main section to make sure it stays unchanged and is not included in the flex.
I updated the photos on the left tab to be smaller to fit the whole screen and make the website fee more coherent and less out of shape.
updated the css and the html to look a bit better and I also fixed a few more issues.